Enable reapply mode
Allows redeployment without providing any input SLAVE_IPS.

Changes in workflow:
* If you set env REAPPLY=yes and ADMIN_IP, fuel_devops is skipped
* If inventory exists and SLAVE_IPS is unset, it retrives SLAVE_IPS
  from inventory.
* Deploy fails if REAPPLY is set, but ADMIN_IP is unset.

Changes to inventory.py:
* Added new functions help, print_cfg, print_ips
* Default to print help if invoked with no args

+++ utils/jenkins/kargo_deploy.sh
@@ -101,7 +101,7 @@ function wait_for_nodes {
 mkdir -p tmp logs
 
 # Allow non-Jenkins script to predefine info
-if [[ -z "$SLAVE_IPS" && -z "$ADMIN_IP" ]]; then
+if [[ -z "$REAPPLY" && -z "$SLAVE_IPS" && -z "$ADMIN_IP" ]]; then
     ENV_TYPE="fuel-devops"
     dos.py erase ${ENV_NAME} || true
     rm -rf logs/*
@@ -112,7 +112,7 @@ if [[ -z "$SLAVE_IPS" && -z "$ADMIN_IP" ]]; then
     ADMIN_IP=${SLAVE_IPS[0]}
     wait_for_nodes $ADMIN_IP
 else
-    ENV_TYPE=${ENV_TYPE:-other}
+    ENV_TYPE=${ENV_TYPE:-other_or_reapply}
     SLAVE_IPS=( $SLAVE_IPS )
     ADMIN_IP=${ADMIN_IP:-${SLAVE_IPS[0]}}
 fi
@@ -191,8 +191,18 @@ if [ -n "$CUSTOM_YAML" ]; then
     custom_opts="-e @$ADMIN_WORKSPACE/kargo/custom.yaml"
 fi
 
-echo "Generating ansible inventory on admin node..."
-admin_node_command CONFIG_FILE=$ADMIN_WORKSPACE/kargo/inventory/inventory.cfg python3 $ADMIN_WORKSPACE/utils/kargo/inventory.py ${SLAVE_IPS[@]}
+# Try to get IPs from inventory if it isn't provided
+if [[ -z "$SLAVE_IPS" ]]; then
+    if admin_node_command stat $ADMIN_WORKSPACE/kargo/inventory/inventory.cfg; then
+        SLAVE_IPS=($(admin_node_command CONFIG_FILE=$ADMIN_WORKSPACE/kargo/inventory/inventory.cfg python3 $ADMIN_WORKSPACE/utils/kargo/inventory.py print_ips))
+    else
+        echo "No slave nodes available. Unable to proceed!"
+        exit_gracefully 1
+    fi
+else
+    echo "Generating ansible inventory on admin node..."
+    admin_node_command CONFIG_FILE=$ADMIN_WORKSPACE/kargo/inventory/inventory.cfg python3 $ADMIN_WORKSPACE/utils/kargo/inventory.py ${SLAVE_IPS[@]}
+fi
 
 echo "Waiting for all nodes to be reachable by SSH..."
 wait_for_nodes ${SLAVE_IPS[@]}
@@ -290,5 +300,6 @@ set -x
     rm -f VLAN_IPS
 fi
 
-
+# TODO(mattymo): Shift to FORCE_NEW instead of REAPPLY
+echo "To reapply deployment, run env REAPPLY=yes ADMIN_IP=$ADMIN_IP $0"
 exit_gracefully ${deploy_res}
